refactor: move logic to interpolation function to own file
